Add utilities to perform some stronghold actions on functions
This is useful if you want to doing something like mark a view public based on
some view_processing middleware of your own without decorating the view
function public and returning a HTTP Response immediately in your middleware.
Also lets me tinker with variable names and such as necessecary without
breaking clients who are doing this. That is a good thing.