Update tests for Django 1.10
In Django 1.10, the user.is_authenticated can act as a method or a
property.  Django depricated its use as a method and will be removed in
Django 2.0.  Django has internally changed its code to use it as a
property.  This patch updates the mock setup of is_authenticated to work
like a property or a method.